Started Working on UI Design
I'd like to know what theme we are going with when it comes to the app, that way I can plan the UI around it. So we just gotta get the ball rolling on what graphics we want.
Updated Part 1, Part 2, Question 1 and Question 2. Getting some feedback from a UI group tomorrow for just button placement.